adding a new script to easily setup the cross comipiler
#!/bin/sh is the interpreter that tells the system to use 'sh' to interpret and execute the commands in the script; more portable.

SCRIPT_DIR stores the path we are currently, which is GCC-cross-compiler-creation and export PATH just add the SCRIPT_DIR with /cross/bin. $PATH is an envoriment variable that contains a colon-seperated list of directories that the shell searches to find the executable program. By colon-seperated list meant that the list contains directories that is seperated by ':' .